Fix reconnect behavior.
In advance a user is allowed to join a parte whether he was
previously in the party or when the party hasn't been closed yet.
Additional all parties are started if hasn't been closed yet
when a new user joins. Plus a new container(channel-master) as added
and it is uncharged of starting parties at start up.